@@ -1,28 +1,33 @@
 name: google_maps_place_picker_mb
 description: A Flutter plugin which provides 'Picking Place' using Google Maps
   widget. (MB version)
-version: 3.0.2
-homepage: https://github.com/martin-braun/google_maps_place_picker_mb
+version: 3.1.0
+homepage: https://github.com/martin-braun-net/google_maps_place_picker_mb
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
-  flutter: ">=3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
   geolocator: ^9.0.2
-  google_api_headers: ^1.5.0+1
-  google_maps_flutter: ^2.2.6
+  google_api_headers: ^2.0.0
+  google_maps_flutter: ^2.3.0
   google_maps_webservice: ^0.0.20-nullsafety.5
-  http: ^0.13.5
+  http: ^1.1.0
   provider: ^6.0.5
-  tuple: ^2.0.1
+  tuple: ^2.0.2
   uuid: ^3.0.7
+  package_info_plus: ^3.1.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
+dependency_overrides:
+  http: any
+  package_info_plus: any
 
 flutter: null
 
